+++ CHANGELOG.md
@@ -6,8 +6,12 @@
 - PowerPoint2007 : Text is subscripted when set superscript to false - @qmachard GH-360
 
 ### Changes
+- PHP 7.1 is now supported - @Progi1984 GH-355
 
 ### Features
+- ODPresentation Writer : Support for the position of Legend - @Progi1984 GH-355
+- ODPresentation Writer : Support for DoughnutChart - @Progi1984 GH-355
+- PowerPoint2007 Writer : Support for DoughnutChart - @Progi1984 GH-355
 
 ## 0.9.0 - 2017-07-05