+++ CHANGELOG.md
@@ -7,11 +7,16 @@
 
 ### Changes
 - PHP 7.1 is now supported - @Progi1984 GH-355
+- PhpOffice\PhpPresentation\Style\Color : Define only the transparency - @Progi1984 GH-370 
 
 ### Features
 - ODPresentation Writer : Support for the position of Legend - @Progi1984 GH-355
 - ODPresentation Writer : Support for DoughnutChart - @Progi1984 GH-355
 - PowerPoint2007 Writer : Support for DoughnutChart - @Progi1984 GH-355
+- ODPresentation Reader : Support for fill for image - @Progi1984 GH-370
+- PowerPoint2007 Reader : Support for fill for image - @Progi1984 GH-370
+- ODPresentation Writer : Support for fill for transparent image - @Progi1984 GH-370
+- PowerPoint2007 Writer : Support for fill for transparent image - @JewrassicPark GH-370
 
 ## 0.9.0 - 2017-07-05