+++ CHANGELOG.md
@@ -1,5 +1,24 @@
 # Changelog
 
+## 0.10.0 - WIP
+
+### Bugfix
+- PowerPoint2007 : Text is subscripted when set superscript to false - @qmachard GH-360
+
+### Changes
+- PHP 7.1 is now supported - @Progi1984 GH-355
+- PhpOffice\PhpPresentation\Style\Color : Define only the transparency - @Progi1984 GH-370 
+- PowerPoint2007 Reader : Background Color based on SchemeColor - @Progi1984 GH-397
+
+### Features
+- ODPresentation Writer : Support for the position of Legend - @Progi1984 GH-355
+- ODPresentation Writer : Support for DoughnutChart - @Progi1984 GH-355
+- PowerPoint2007 Writer : Support for DoughnutChart - @Progi1984 GH-355
+- ODPresentation Reader : Support for fill for image - @Progi1984 GH-370
+- PowerPoint2007 Reader : Support for fill for image - @Progi1984 GH-370
+- ODPresentation Writer : Support for fill for transparent image - @Progi1984 GH-370
+- PowerPoint2007 Writer : Support for fill for transparent image - @JewrassicPark GH-370
+
 ## 0.9.0 - 2017-07-05
 
 ### Bugfix